+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [1.1.1] - 2026-03-13
+
+### Added
+
+- Added `eslint-plugin-react-refresh` plugin support
+- Added `eslint-plugin-valtio` plugin support
+
+### Changed
+
+- Import `z` from `@kokimoki/app` instead of `@kokimoki/kit` (zod is now re-exported from app)
+- Update `@kokimoki/app` to `3.1.6`
+- Update `@kokimoki/kit` to `1.8.2`
+- Update `@kokimoki/react-components` to `1.0.2`
+
 ## [1.1.0] - 2026-02-14
 
 ### Added